pyproject.toml
[project]
name = "translate-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.124.0",
    "jinja2>=3.1.6",
    "langchain>=1.1.3",
    "langchain-core>=1.1.2",
    "langchain-openai>=1.1.1",
    "langchain-text-splitters>=1.0.0",
    "langgraph>=1.0.4",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "tiktoken>=0.12.0",
    "uvicorn>=0.38.0",
]
